Add files via upload
In this particular project, we are using a dataset that contains information likeloan id, gender education etc and using that to classify the loan status. 
However, before you go ahead with the project, it is advised that you first pre-process the data, since it may contain some irregularities and noise. In addition, try various tricks and techniques in order to gain the best accuracy in your predictions.